requirements.txt
pytest==7.3.1
pytest-xdist==3.3.1
pytest-html==3.2.0
selenium==4.32.0
webdriver-manager==4.0.2
allure-pytest==2.13.2
allure-python-commons==2.13.2
pytest-rerunfailures==11.1.2
requests==2.32.3
python-dotenv==1.0.0
Faker==19.3.0
# Add specific version for compatibility
urllib3==2.4.0